Big refactor
Essentially the YTStreamRepo now searches based on if you want to include author and lyrics in the query which is passed directly from the UI level via either the IDownloadService:DownloadMediaAsync() or IStreamRepository:GetStreamOfTrackAsync() .

I have also hardcoded it to filter the songs duration to be no more than 30 seconds more or less than the Track entities duration.

We also now have a try/catch block to handle possible 403 errors that can be caused by either ratelimits or when youtube changes stuff and YoutubeExplode isn't updated.

Note: This Console.WriteLine will eventually be turned into a log method. as per #14