Minor Refactor YoutubeDownloadService
This refactor essentially does two things.

The first is to pass the bool for whether the search query should have author or lyrics in it.

The second change is to ensure the tracks save name is formated in a way which is allowed on all operating systems. (Essentially it removes illegal chars and formats the name with an author).