Add SearchAsync to IStreamRepository
This method allows you to search Youtube by directly providing your own Track entity. This will be handy for when you want to provide a simple search funtion in a UI to the user.

Method returns an IEnumerable<Track> with the youtube information such as title, author & duration which you can then use to provide direct download functionality for users.

Ideally you would want to use this to search youtube from a UI, then pass in the selected Track entity into the DownloadMediaAsync method.